redact sensitive information from logs
