+++ viseron/__main__.py
@@ -2,6 +2,7 @@
 import logging
 
 from viseron import Viseron
+from viseron.helpers.log_filter import DuplicateFilter
 
 LOGGER = logging.getLogger()
 
@@ -50,27 +51,6 @@ class MyFormatter(logging.Formatter):
         self._style._fmt = format_orig
 
         return result
-
-
-class DuplicateFilter(logging.Filter):
-    """Formats identical log entries to overwrite the last."""
-
-    # pylint: disable=attribute-defined-outside-init
-    def filter(self, record):
-        current_log = (record.name, record.module, record.levelno, record.msg)
-        try:
-            if current_log != getattr(self, "last_log", None):
-                self.last_log = current_log
-                self.current_count = 0
-            else:
-                self.current_count += 1
-                if self.current_count > 0:
-                    record.msg = "{}, message repeated {} times".format(
-                        record.msg, self.current_count + 1
-                    )
-        except ValueError:
-            pass
-        return True
 
 
 def main():
